requirements.txt
google-adk==1.28.1
fastapi==0.135.3
uvicorn==0.42.0
